Spin SRP status off into its own component
Change SrpLossRow to just embed the component. Allows us to reuse it
for battle reports later on.
